requirements.txt
pymongo==4.6.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0
